requirements.txt
numpy==1.24.3
Flask==2.2.2
streamlit
rembg
requests
Pillow